+++ packages/amplify-console-integration-tests/src/consoleHosting/utils.ts
@@ -1,6 +1,7 @@
 import * as fs from 'fs-extra';
 import * as path from 'path';
 import { spawnSync } from 'child_process';
+import * as util from '../util';
 
 import { HOSTING, RESOURCE, TYPE, TYPE_UNKNOWN, CATEGORIES, APPID, PROVIDER } from './constants';
 
@@ -37,11 +38,12 @@ function readJsonFile(jsonFilePath: string, encoding = 'utf8') {
   return JSON.parse(fs.readFileSync(jsonFilePath, encoding));
 }
 
-export function createTestProject(): string {
-  const projectName = `console-hosting-${Math.round(Math.random() * 100)}-test-${Math.round(Math.random() * 1000)}`;
-  const projectDir = path.normalize(path.join(__dirname, '../../../..'));
+export async function createTestProject(): Promise<string> {
+  const projRoot = await util.createNewProjectDir('console-hosting');
+  const projectName = path.basename(projRoot);
+  const projectDir = path.dirname(projRoot);
 
   spawnSync('npx', ['create-react-app', projectName], { cwd: projectDir });
 
-  return path.join(projectDir, projectName);
+  return projRoot;
 }
