+++ packages/amplify-e2e-core/src/index.ts
@@ -1,7 +1,16 @@
 import path from 'path';
-import { spawnSync } from 'child_process';
+import * as fs from 'fs-extra';
+import { spawnSync, execSync } from 'child_process';
 
 export * from './utils/nexpect';
+
+declare global {
+  namespace NodeJS {
+    interface Global {
+      getRandomId: () => string;
+    }
+  }
+}
 
 export function getCLIPath() {
   if (isCI()) {
@@ -16,4 +25,18 @@ export function isCI(): boolean {
 
 export function npmInstall(cwd: string) {
   spawnSync('npm', ['install'], { cwd });
+}
+
+export async function createNewProjectDir(projectName: string, prefix = path.join('/tmp', 'amplify-e2e-tests')): Promise<string> {
+  const currentHash = execSync('git rev-parse --short HEAD', { cwd: __dirname })
+    .toString()
+    .trim();
+  let projectDir;
+  do {
+    const randomId = await global.getRandomId();
+    projectDir = path.join(prefix, `${projectName}_${currentHash}_${randomId}`);
+  } while (fs.existsSync(projectDir));
+
+  fs.ensureDirSync(projectDir);
+  return projectDir;
 }
